requirements.txt
discord.py[docs]
Flask
python-dotenv
